+++ test/unit/range/views/CMakeLists.txt
@@ -14,7 +14,6 @@ seqan3_test(move_test.cpp)
 seqan3_test(persist_test.cpp)
 seqan3_test(rank_to_test.cpp)
 seqan3_test(repeat_n_test.cpp)
-seqan3_test(repeat_test.cpp)
 seqan3_test(type_reduce_test.cpp)
 seqan3_test(slice_test.cpp)
 seqan3_test(take_test.cpp)
