+++ test/unit/utility/views/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_subdirectories()
 
+seqan3_test(repeat_test.cpp)
 seqan3_test (zip_test.cpp)
