+++ test/unit/utility/views/repeat_test.cpp
@@ -13,11 +13,11 @@
 #include <seqan3/core/detail/iterator_traits.hpp>
 #include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/range/views/persist.hpp>
-#include <seqan3/range/views/repeat.hpp>
 #include <seqan3/range/views/take.hpp>
 #include <seqan3/range/views/take_exactly.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
 #include <seqan3/test/expect_same_type.hpp>
+#include <seqan3/utility/views/repeat.hpp>
 
 TEST(repeat_view, deduction_guide)
 {
